pyproject.toml
[project]
name = "messenger-answering-machine"
version = "0.1.0"
description = "AI-Powered Messenger-Anrufbeantworter – Sprachgesteuerte Zusammenfassung ungelesener Nachrichten"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "anthropic>=0.40.0",
    "deepgram-sdk>=3.7.0",
    "edge-tts>=6.1.0",
    "twilio>=9.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydub>=0.25.0",
    "audioop-lts>=0.2.0",
    "sounddevice>=0.4.0",
    "numpy>=1.26.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.6.0",
    "tzdata>=2025.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["app"]
